Update Thrown.md
We do not need a throwing strike action. We can use PF2e's Strike action and Thrown trait interaction as it would make no difference and simply complicate the rules. Sure there'll be bombs that are thrown but those are still attacks either way.

I may be wrong here, so feel free to thoroughly correct me if that's the case.